quick and dirty 1.20 port
A very mechanical port to 1.20 based on suggestions from the Fabric team
here https://fabricmc.net/2023/05/25/120.html. We did it to allow us to
start testing for a server update to 1.20. Feel free to use it any way
you want.

Mainly it is updating calls to sendFeedback() to pass a Supplier<Text>
instead of a Text parameter.

The changes to handle the removal of Material are not fully tested. The
isSolid() call is marked as deprecated, so that should probably be
something else.